enhance(backend): センシティブワードの設定がハッシュタグトレンドにも適用されるように
