+++ CHANGELOG.md
@@ -12,6 +12,17 @@
 
 -->
 
+## 2023.12.1
+
+### General
+-
+
+### Client
+- 
+
+### Server
+- Enhance: センシティブワードの設定がハッシュタグトレンドにも適用されるようになりました
+
 ## 2023.12.0
 
 ### Note
